backend_app/requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.11
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
psycopg2-binary==2.9
pytz==2021.3
sqlparse==0.4.2
gunicorn==20.1.0
django-mptt==0.13.4
django-mptt-admin==2.2.0
django-cors-headers==3.11.0
django-filter==21.1